handwriting_ocr/requirements.txt
# FastAPI Application Dependencies

# Core Web Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6

# ML/DL Framework
torch==2.1.0
torchvision==0.16.0
ultralytics==8.0.200
transformers==4.35.0
accelerate==0.24.0

# Image Processing
Pillow==10.1.0
opencv-python-headless==4.8.1.78

# Utilities
numpy==1.24.3
pydantic==2.4.2
python-dotenv==1.0.0

# AWS SDK (optional)
boto3==1.29.0

# Logging & Monitoring
python-json-logger==2.0.7
